Backend: Create Message event handlers
