+++ .github/workflows/pylint.yml
@@ -14,11 +14,11 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Python flake8 Lint
+      uses: py-actions/flake8@v2.2.1
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt 
+        pip install
         pip install pylint pylint-flask pylint-flask-sqlalchemy
-    - name: Analysing the code with pylint
-      run: |
-        pylint $(git ls-files '*.py') --load-plugins pylint_flask_sqlalchemy,pylint_flask --init-hook "from pylint.config import find_pylintrc; import sys; sys.path.append(find_pylintrc())"
+
