+++ backend/db/model.py
@@ -1,6 +1,16 @@
 """defines data schema"""
-from sqlalchemy import Column, Integer, String, ForeignKey, Table, Boolean, DateTime
+from sqlalchemy import (
+    Boolean,
+    Column,
+    DateTime,
+    ForeignKey,
+    Integer,
+    String,
+    Table
+)
 from sqlalchemy.orm import declarative_base, relationship
+from sqlalchemy.dialects.postgresql import UUID
+import uuid
 
 Base = declarative_base()
 
@@ -34,7 +44,11 @@ class User(Base):
     )
 
     def __repr__(self):
-        return f"User(id={self.id!r}, name={self.username!r}, password={self.password!r})"
+        return (
+            f"User(id={self.id!r}, "
+            f"name={self.username!r}, "
+            f"password={self.password!r})"
+        )
 
     def as_dict(self):
         """
@@ -55,6 +69,7 @@ class Chatroom(Base):
     __tablename__ = 'chatrooms'
 
     id = Column(Integer, primary_key=True)
+    public_id = Column(UUID(as_uuid=True), default=uuid.uuid4)
     name = Column(String(512), nullable=False)
     owner_id = Column(Integer, ForeignKey("users.id"), nullable=False)
 
@@ -69,7 +84,23 @@ class Chatroom(Base):
     )
 
     def __repr__(self) -> str:
-        return f"Chatroom(id={self.id!r}, name={self.name!r}, owner={self.owner.username!r})"
+        return (
+            f"Chatroom(public_id={self.public_id!r}, "
+            f"name={self.name!r}, "
+            f"owner={self.owner.username!r})"
+        )
+
+    def as_dict(self):
+        """
+        Returns the chatroom parameters as a dictionary
+        """
+        return {
+            'name': self.name,
+            'owner': self.owner.as_dict(),
+            'users': [user.as_dict() for user in self.users],
+            'public_ID': str(self.public_id),
+            'messages': [m.as_dict() for m in self.messages]
+        }
 
 
 class Message(Base):
@@ -88,6 +119,18 @@ class Message(Base):
 
     def __repr__(self) -> str:
         return f"Message(id={self.id!r})"
+
+    def as_dict(self):
+        """
+        Returns the chatroom parameters as a dictionary
+        """
+        return {
+            'id': self.id,
+            'sent': self.sent.isoformat(),
+            'owner': self.from_user.username,
+            'chatroom': str(self.chatroom.public_id),
+            'content': self.content
+        }
 
 
 Base.registry.configure()
